requirements.txt
cryptography==40.0.1
Django==5.1.1
djangorestframework==3.15.0

